backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
pydantic-settings==2.4.0
aiofiles==24.1.0
aiohttp==3.10.5
httpx==0.27.2
openai>=1.45.0
moviepy==1.0.3
Pillow==10.4.0
numpy==1.26.4
python-dotenv==1.0.1
websockets==12.0
PyYAML==6.0.2
